formatted the code
